requirements.txt
python-telegram-bot==20.3
Flask==3.1.0
gspread==6.2.1
google-auth==2.41.1
google-auth-oauthlib==1.2.2

